Grouping opt's http.Request implementations with the Route
