+++ opt/server_capability.go
@@ -1,0 +1,10 @@
+package opt
+
+import "io"
+
+type StaticCapabilityController struct {
+}
+
+func (controller *StaticCapabilityController) Options(writer io.Writer) {
+	panic("implement me")
+}
